Added access token implementation
Finished createWebToken function and worked on the user login api. All thats left is to 1) make sure passwords are hashed before being sent to the backend so that no plaintext is stored in database, and implement secure http only cookies for refresh tokens so that when new access tokens are requested we can compare this hashed http token to the stored refresh token in the db to verify. Then can work on adding users and doing more frontend stuff